requirements.txt
# Core dependencies
click==8.1.7
pyyaml==6.0.1
pydub==0.25.1
srt==3.5.3

# TTS services
google-cloud-texttospeech==2.14.1
azure-cognitiveservices-speech==1.34.0

# Progress display
rich==13.7.0

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0

# Development dependencies (optional, install with pip install -r requirements-dev.txt)
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.12.1
# flake8==7.0.0
# mypy==1.8.0